+ setup project to use blueprints and moved user authentication into new auth file
