+++ config.py
@@ -1,0 +1,3 @@
+class Config(object):
+    SQLALCHEMY_DATABASE_URI = 'sqlite:///should_i_buy_it.db'
+    SECRET_KEY = 'secret_key'
