+++ database/models.py
@@ -11,11 +11,11 @@ class User(db.Model):
     password = db.Column(db.String(255), nullable=False)
     profile_picture = db.Column(db.String(255))
 
+    session = db.relationship('Session', backref='user', lazy=True)
     images = db.relationship('Image', backref='user', lazy=True)
     votes = db.relationship('Vote', backref='user', lazy=True)
     comments = db.relationship('Comment', backref='user', lazy=True)
     followers = db.relationship('Follow', foreign_keys='Follow.follower_id', backref='follower', lazy=True)
-    followees = db.relationship('Follow', foreign_keys='Follow.followee_id', backref='followee', lazy=True)
 
 # images table metadata
 class Image(db.Model):
@@ -47,3 +47,11 @@ class Follow(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     follower_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
     followee_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
+
+# store session data
+class Session(db.Model):
+
+    #also store the hashed id in the browser's localstorage for referencing
+    id = db.Column(db.LargeBinary, primary_key=True)
+    session_time = db.Column(db.DateTime, default=datetime.utcnow)
+    user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
