+++ static/js/registerUser.js
@@ -19,7 +19,7 @@ $(document).ready(function() {
         // Send the data using AJAX
         $.ajax({
             type: 'POST',
-            url: 'http://localhost:5000/register',
+            url: 'http://localhost:5000/auth/register',
             data: formData,
             success: function(response) {
                 // Handle success
